move helpers/ to plugin_helpers/
Why:

* Comply with wazo-notebook/file-structure.md
