+++ wazo_calld/plugin_helpers/exceptions.py
