+++ wazo_calld/plugins/applications/services.py
@@ -6,9 +6,8 @@ import time
 
 from requests import HTTPError
 from ari.exceptions import ARINotFound
-from wazo_calld.helpers import ami
-from wazo_calld.helpers import confd
-from wazo_calld.helpers.ari_ import Channel
+from wazo_calld.plugin_helpers import ami, confd
+from wazo_calld.plugin_helpers.ari_ import Channel
 from wazo_calld.exceptions import InvalidExtension
 from .models import (
     CallFormatter,
