+++ wazo_calld/plugins/calls/bus_consume.py
@@ -10,8 +10,8 @@ from xivo_bus.resources.calls.hold import CallOnHoldEvent
 from xivo_bus.resources.calls.hold import CallResumeEvent
 from xivo_bus.resources.common.event import ArbitraryEvent
 
-from wazo_calld.helpers import ami
-from wazo_calld.helpers.ari_ import Channel
+from wazo_calld.plugin_helpers import ami
+from wazo_calld.plugin_helpers.ari_ import Channel
 
 from .schema import call_schema
