+++ wazo_calld/plugins/conferences/services.py
@@ -6,8 +6,8 @@ import logging
 from ari.exceptions import ARINotFound
 from marshmallow import ValidationError
 from requests import RequestException
-from wazo_calld.helpers.confd import Conference
-from wazo_calld.helpers.ari_ import Channel
+from wazo_calld.plugin_helpers.confd import Conference
+from wazo_calld.plugin_helpers.ari_ import Channel
 from wazo_calld.exceptions import (
     WazoAmidError,
 )
