+++ wazo_calld/plugins/faxes/services.py
@@ -8,8 +8,8 @@ import subprocess
 
 from tempfile import mkstemp
 
-from wazo_calld.helpers import ami
-from wazo_calld.helpers.confd import User
+from wazo_calld.plugin_helpers import ami
+from wazo_calld.plugin_helpers.confd import User
 from wazo_calld.exceptions import InvalidExtension
 
 from .exceptions import FaxFailure
