+++ wazo_calld/plugins/relocates/services.py
@@ -6,13 +6,13 @@ import logging
 import threading
 
 from wazo_calld.exceptions import UserPermissionDenied
-from wazo_calld.helpers import ami
-from wazo_calld.helpers.ari_ import (
+from wazo_calld.plugin_helpers import ami
+from wazo_calld.plugin_helpers.ari_ import (
     ARINotFound,
     Channel,
 )
-from wazo_calld.helpers.confd import User
-from wazo_calld.helpers.exceptions import (
+from wazo_calld.plugin_helpers.confd import User
+from wazo_calld.plugin_helpers.exceptions import (
     NotEnoughChannels,
     TooManyChannels,
     InvalidUserLine,
