+++ wazo_calld/plugins/relocates/state.py
@@ -11,8 +11,8 @@ from ari.exceptions import (
 )
 from wazo_calld.ari_ import DEFAULT_APPLICATION_NAME
 from wazo_calld.exceptions import WazoAmidError
-from wazo_calld.helpers import ami
-from wazo_calld.helpers.ari_ import Channel
+from wazo_calld.plugin_helpers import ami
+from wazo_calld.plugin_helpers.ari_ import Channel
 
 from .exceptions import (
     RelocateCreationError,
