+++ wazo_calld/plugins/switchboards/confd.py
@@ -5,7 +5,7 @@ from requests import HTTPError
 from requests import RequestException
 
 from wazo_calld.exceptions import WazoConfdUnreachable
-from wazo_calld.helpers.confd import not_found
+from wazo_calld.plugin_helpers.confd import not_found
 
 
 class Switchboard:
