+++ wazo_calld/plugins/transfers/ari_helpers.py
@@ -7,7 +7,7 @@ import logging
 from ari.exceptions import ARINotFound, ARINotInStasis
 
 from wazo_calld.exceptions import WazoAmidError
-from wazo_calld.helpers import ami
+from wazo_calld.plugin_helpers import ami
 
 logger = logging.getLogger(__name__)
