+++ wazo_calld/plugins/transfers/resources.py
@@ -8,7 +8,7 @@ from marshmallow.validate import OneOf, Length
 from wazo_calld.auth import get_token_user_uuid_from_request
 from wazo_calld.auth import required_acl
 from wazo_calld.http import AuthResource
-from wazo_calld.helpers.mallow import StrictDict
+from wazo_calld.plugin_helpers.mallow import StrictDict
 
 
 class TransferRequestSchema(Schema):
