+++ wazo_calld/plugins/transfers/services.py
@@ -8,12 +8,12 @@ from xivo.caller_id import assemble_caller_id
 from wazo_calld.ari_ import DEFAULT_APPLICATION_NAME
 from wazo_calld.exceptions import UserPermissionDenied
 from wazo_calld.exceptions import InvalidExtension
-from wazo_calld.helpers.ari_ import Channel
-from wazo_calld.helpers.confd import User
-from wazo_calld.helpers.exceptions import NotEnoughChannels
-from wazo_calld.helpers.exceptions import TooManyChannels
+from wazo_calld.plugin_helpers.ari_ import Channel
+from wazo_calld.plugin_helpers.confd import User
+from wazo_calld.plugin_helpers.exceptions import NotEnoughChannels
+from wazo_calld.plugin_helpers.exceptions import TooManyChannels
 
-from wazo_calld.helpers import ami
+from wazo_calld.plugin_helpers import ami
 from . import ari_helpers
 from .exceptions import NoSuchTransfer
 from .exceptions import TooManyTransferredCandidates
