+++ wazo_calld/plugins/transfers/state.py
@@ -6,7 +6,7 @@ import uuid
 
 from ari.exceptions import ARINotFound
 
-from wazo_calld.helpers.ari_ import Channel
+from wazo_calld.plugin_helpers.ari_ import Channel
 from . import ari_helpers
 from .exceptions import TransferAnswerError
 from .exceptions import TransferCreationError
