+++ wazo_calld/plugins/transfers/state_persistor.py
@@ -1,12 +1,14 @@
-# Copyright 2016-2017 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
 
-from wazo_calld.helpers.ari_ import (GlobalVariableAdapter,
-                                     GlobalVariableJsonAdapter,
-                                     GlobalVariableNameDecorator,
-                                     GlobalVariableConstantNameAdapter)
+from wazo_calld.plugin_helpers.ari_ import (
+    GlobalVariableAdapter,
+    GlobalVariableJsonAdapter,
+    GlobalVariableNameDecorator,
+    GlobalVariableConstantNameAdapter,
+)
 from .transfer import Transfer
 
 logger = logging.getLogger(__name__)
